[CHORE] Re-organised React app structure.
