+++ src/components/FuelSavingsTextInput.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+/**
+ * FuelSavingsTextInput Component
+ */
 const FuelSavingsTextInput = ({name, value, placeholder, onChange}) => {
   return (
     <input
@@ -16,9 +19,21 @@ const FuelSavingsTextInput = ({name, value, placeholder, onChange}) => {
 const { string, func, number, oneOfType } = PropTypes;
 
 FuelSavingsTextInput.propTypes = {
+  /**
+   * Holds the name for the input
+   * */
   name: string.isRequired,
+  /**
+   * Holds the change event handler for the input
+   * */
   onChange: func.isRequired,
+  /**
+   * Holds the placeholder for the input
+   * */
   placeholder: string,
+  /**
+   * Holds the value for the input
+   * */
   value: oneOfType([
     string,
     number
