+++ src/components/FuelSavingsTextInput.md
@@ -1,0 +1,3 @@
+```js
+<FuelSavingsTextInput name="saving" placeholder="Savings" value="" onChange={() => {}}></FuelSavingsTextInput>
+```
