+++ src/containers/Root.jsx
@@ -2,7 +2,8 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { ConnectedRouter } from 'react-router-redux';
 import { Provider } from 'react-redux';
-import App from './App';
+
+import App from '../screens/App';
 
 export default class Root extends Component {
   render() {
