+++ src/screens/AboutPage.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+
 import '../styles/about-page.css';
 
 // Since this component is simple and static, there's no parent container for it.
