+++ src/screens/FuelSavingsPage.js
@@ -2,8 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
-import * as actions from '../../actions/fuelSavingsActions';
-import FuelSavingsForm from '../FuelSavingsForm';
+
+import * as actions from '../actions/fuelSavingsActions';
+import FuelSavingsForm from '../components/FuelSavingsForm';
 
 export class FuelSavingsPage extends React.Component {
   saveFuelSavings = () => {
